fix: remove restricted emulation as we now have DOM isolation
